feat(recipeDetail): 레시피 상세 조회 네비게이팅 구현

+++ Android/core/data-api/src/main/java/com/sundaegukbap/banchango/core/data/repository/api/RecipeRepository.kt
@@ -4,5 +4,5 @@ import com.sundaegukbap.banchango.Recipe
 
 interface RecipeRepository {
     suspend fun getRecipeRecommendation(): Result<List<Recipe>>
-    suspend fun getRecipeDetail(id: Int): Result<Recipe>
+    suspend fun getRecipeDetail(id: Long): Result<Recipe>
 }
